+++ src/test/java/com/gaming_platform/GamingPlatformApplicationTests.java
@@ -1,4 +1,4 @@
-package com.gamingPlatform;
+package com.gaming_platform;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
